Improved value precision for percents in sizes.
git-svn-id: https://svn.ctdp.net/svn/rfdynhud/trunk@211 1c0d6ce8-a1e3-b74c-a49d-f790e2421cd5

+++ java/src/net/ctdp/rfdynhud/widgets/widget/Widget.java
@@ -136,7 +136,7 @@ public abstract class Widget implements Documented
      * @param newWidth
      * @param newHeight
      */
-    protected void onSizeChanged( float oldWidth, float oldHeight, float newWidth, float newHeight )
+    protected void onSizeChanged( int oldWidth, int oldHeight, int newWidth, int newHeight )
     {
         WidgetsConfiguration wc = getConfiguration();
